More test coverage.
